+++ src/pdi_migration/validator/__init__.py
@@ -1,11 +1,14 @@
 from pdi_migration.validator.assess import assess_source
+from pdi_migration.validator.effort import EffortEstimate, build_effort
 from pdi_migration.validator.gaps import GapReport, build_gap_report
 from pdi_migration.validator.impact import ImpactAnalysis, build_impact_analysis
 from pdi_migration.validator.report import MigrationReport, build_report
 from pdi_migration.validator.score import MigrationScore, build_score
 
 __all__ = [
+    "EffortEstimate",
     "GapReport",
+    "build_effort",
     "ImpactAnalysis",
     "MigrationReport",
     "MigrationScore",